cleanup and rename files
