+++ data/hackathon2025/agenda.json
