+++ data/hackathon2025/participants.json
