+++ data/hackathon2025/sponsors.json
